mock_weather_api/kafka_producer.py
import json
from kafka import KafkaProducer

producer = KafkaProducer(
    bootstrap_servers="localhost:9092",
    value_serializer=lambda v: json.dumps(v).encode("utf-8"),
    linger_ms=10,
    batch_size=16384
)

def send_to_kafka(topic, data):
    producer.send(topic, data)